change: rename CreateGroup to ManageGroup
