--- src/pages/CreateGroup/index.ts
@@ -1,5 +1,0 @@
-import { CreateGroupPage } from "./CreateGroup";
-
-export default CreateGroupPage;
-export * from "./CreateGroup";
-export * from "./loader";
