+++ src/pages/ManageGroup/index.ts
@@ -1,0 +1,5 @@
+import { ManageGroupPage } from "./ManageGroup";
+
+export default ManageGroupPage;
+export * from "./ManageGroup";
+export * from "./loader";
